Refactor frontend: Add dependencies, update routing, and enhance project structure
