+++ frontend/package.json
@@ -12,7 +12,18 @@
   "dependencies": {
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
-    "react-router-dom": "^7.8.0"
+    "react-router-dom": "^7.8.0",
+    "axios": "^1.7.9",
+    "chart.js": "^4.4.4",
+    "react-chartjs-2": "^5.2.0",
+    "framer-motion": "^11.3.31",
+    "react-hot-toast": "^2.5.1",
+    "react-hook-form": "^7.53.0",
+    "zustand": "^5.0.1",
+    "lucide-react": "^0.468.0",
+    "date-fns": "^3.6.0",
+    "@headlessui/react": "^2.2.0",
+    "clsx": "^2.1.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.32.0",
