--- index.html
@@ -1,14 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 24 24' fill='%2316a085'><path d='M16 11V7a4 4 0 00-8 0v4H6v9h12v-9h-2zm-6-4a2 2 0 114 0v4h-4V7z'/></svg>" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta name="description" content="Professional e-commerce data automation platform. Sync inventory, pricing, and orders across all your sales channels." />
-    <title>E-commerce Data Automation - Premium Dashboard</title>
-  </head>
-  <body>
-    <div id="root"></div>
-    <script type="module" src="/src/main.jsx"></script>
-  </body>
-</html>
