--- src/index.css
@@ -1,172 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@layer base {
-  html {
-    scroll-behavior: smooth;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    text-rendering: optimizeLegibility;
-  }
-
-  body {
-    @apply bg-gradient-to-br from-primary-50 to-info-50 min-h-screen;
-  }
-}
-
-@layer components {
-  /* Glassmorphism Effect */
-  .glass {
-    @apply bg-white/25 backdrop-blur-lg border border-white/20;
-  }
-
-  .glass-card {
-    @apply glass rounded-3xl shadow-2xl;
-  }
-
-  /* Premium Button Styles */
-  .btn-primary {
-    @apply inline-flex items-center gap-2 px-6 py-3 rounded-xl font-semibold text-white 
-           bg-gradient-to-r from-primary-500 to-primary-600 
-           hover:from-primary-600 hover:to-primary-700
-           transform hover:-translate-y-0.5 hover:shadow-lg hover:shadow-primary-500/40
-           transition-all duration-300 cursor-pointer relative overflow-hidden;
-  }
-
-  .btn-secondary {
-    @apply inline-flex items-center gap-2 px-6 py-3 rounded-xl font-semibold 
-           bg-white/80 text-neutral-700 border border-white/30 backdrop-blur-sm
-           hover:bg-white/95 hover:-translate-y-0.5 hover:shadow-lg
-           transition-all duration-300 cursor-pointer;
-  }
-
-  .btn-glow::before {
-    content: '';
-    @apply absolute top-0 left-[-100%] w-full h-full 
-           bg-gradient-to-r from-transparent via-white/40 to-transparent
-           transition-all duration-500;
-  }
-
-  .btn-glow:hover::before {
-    @apply left-full;
-  }
-
-  /* Card Styles */
-  .card {
-    @apply bg-white rounded-2xl p-6 shadow-lg border border-neutral-200 
-           hover:shadow-xl hover:-translate-y-1 transition-all duration-300;
-  }
-
-  .kpi-card {
-    @apply card relative overflow-hidden;
-    min-height: 200px;
-  }
-
-  .kpi-card::before {
-    content: '';
-    @apply absolute -top-1/2 -right-1/2 w-full h-full rounded-full opacity-5;
-  }
-
-  .kpi-card.revenue::before {
-    @apply bg-success-500;
-  }
-
-  .kpi-card.orders::before {
-    @apply bg-info-500;
-  }
-
-  .kpi-card.inventory::before {
-    @apply bg-warning-500;
-  }
-
-  .kpi-card.platforms::before {
-    @apply bg-primary-500;
-  }
-
-  /* Loading Animation */
-  .loading-ring {
-    @apply border-4 border-transparent border-t-primary-500 rounded-full animate-spin;
-  }
-
-  .loading-ring:nth-child(2) {
-    @apply border-t-success-500 animate-spin;
-    animation-delay: -0.5s;
-  }
-
-  .loading-ring:nth-child(3) {
-    @apply border-t-warning-500 animate-spin;
-    animation-delay: -1s;
-  }
-
-  /* Activity Items */
-  .activity-item {
-    @apply flex items-center gap-4 p-4 rounded-xl border-l-4 
-           hover:bg-neutral-50 transition-all duration-200;
-  }
-
-  .activity-item.success {
-    @apply border-l-success-500;
-  }
-
-  .activity-item.warning {
-    @apply border-l-warning-500;
-  }
-
-  .activity-item.error {
-    @apply border-l-error-500;
-  }
-
-  .activity-item.info {
-    @apply border-l-info-500;
-  }
-}
-
-@layer utilities {
-  .text-gradient {
-    @apply bg-gradient-to-r from-primary-500 to-success-500 bg-clip-text text-transparent;
-  }
-
-  .animate-entrance {
-    @apply animate-slide-up;
-  }
-
-  .floating-shape {
-    @apply absolute rounded-full opacity-10 animate-float;
-  }
-
-  .floating-shape-1 {
-    width: 3.75rem;
-    height: 3.75rem;
-    top: 25%;
-    left: 8.333%;
-    @apply bg-primary-500;
-  }
-
-  .floating-shape-2 {
-    width: 5rem;
-    height: 5rem;
-    top: 50%;
-    right: 16.666%;
-    @apply bg-success-500;
-    animation-delay: -5s;
-  }
-
-  .floating-shape-3 {
-    width: 2.5rem;
-    height: 2.5rem;
-    bottom: 33.333%;
-    left: 60%;
-    @apply bg-info-500;
-    animation-delay: -10s;
-  }
-
-  .floating-shape-4 {
-    width: 1.5rem;
-    height: 1.5rem;
-    top: 16.666%;
-    right: 50%;
-    @apply bg-warning-500;
-    animation-delay: -15s;
-  }
-}
