--- tailwind.config.js
@@ -1,107 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-export default {
-  content: [
-    "./index.html",
-    "./src/**/*.{js,ts,jsx,tsx}",
-  ],
-  theme: {
-    extend: {
-      colors: {
-        primary: {
-          50: '#f0fdfa',
-          100: '#ccfbf1',
-          400: '#2dd4bf',
-          500: '#16a085',
-          600: '#138d75',
-          700: '#0d9488',
-        },
-        success: {
-          50: '#f0fdf4',
-          100: '#dcfce7',
-          400: '#4ade80',
-          500: '#22c55e',
-          600: '#16a34a',
-        },
-        warning: {
-          50: '#fffbeb',
-          100: '#fef3c7',
-          400: '#fbbf24',
-          500: '#f59e0b',
-          600: '#d97706',
-        },
-        error: {
-          50: '#fef2f2',
-          100: '#fee2e2',
-          400: '#f87171',
-          500: '#ef4444',
-          600: '#dc2626',
-        },
-        info: {
-          50: '#eff6ff',
-          100: '#dbeafe',
-          400: '#60a5fa',
-          500: '#3b82f6',
-          600: '#2563eb',
-        },
-        neutral: {
-          50: '#f8fafc',
-          100: '#f1f5f9',
-          200: '#e2e8f0',
-          300: '#cbd5e1',
-          400: '#94a3b8',
-          500: '#64748b',
-          600: '#475569',
-          700: '#334155',
-          800: '#1e293b',
-          900: '#0f172a',
-          950: '#020617',
-        }
-      },
-      fontFamily: {
-        sans: ['-apple-system', 'BlinkMacSystemFont', 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'sans-serif'],
-      },
-      animation: {
-        'float': 'float 20s ease-in-out infinite',
-        'fade-in': 'fadeIn 0.6s ease-out',
-        'slide-up': 'slideUp 0.6s cubic-bezier(0.4, 0, 0.2, 1)',
-        'scale-in': 'scaleIn 0.6s cubic-bezier(0.68, -0.55, 0.265, 1.55)',
-        'spin-slow': 'spin 3s linear infinite',
-        'pulse-glow': 'pulseGlow 2s ease-in-out infinite',
-      },
-      keyframes: {
-        float: {
-          '0%, 100%': { transform: 'translate(0, 0) rotate(0deg)' },
-          '25%': { transform: 'translate(30px, -30px) rotate(90deg)' },
-          '50%': { transform: 'translate(-20px, 20px) rotate(180deg)' },
-          '75%': { transform: 'translate(-30px, -10px) rotate(270deg)' },
-        },
-        fadeIn: {
-          '0%': { opacity: '0', transform: 'translateY(30px)' },
-          '100%': { opacity: '1', transform: 'translateY(0)' },
-        },
-        slideUp: {
-          '0%': { opacity: '0', transform: 'translateY(30px)' },
-          '100%': { opacity: '1', transform: 'translateY(0)' },
-        },
-        scaleIn: {
-          '0%': { transform: 'scale(0)' },
-          '50%': { transform: 'scale(1.1)' },
-          '100%': { transform: 'scale(1)' },
-        },
-        pulseGlow: {
-          '0%, 100%': { boxShadow: '0 0 20px rgba(22, 160, 133, 0.4)' },
-          '50%': { boxShadow: '0 0 30px rgba(22, 160, 133, 0.6)' },
-        },
-      },
-      backdropBlur: {
-        xs: '2px',
-      },
-      spacing: {
-        '15': '3.75rem',
-        '18': '4.5rem',
-        '88': '22rem',
-      }
-    },
-  },
-  plugins: [],
-}
